fix: deploy args
