+++ src/components/Deploy/3-ReviewStep/ReviewStep.tsx
@@ -1,25 +1,37 @@
+import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/24/solid'
 import { Button } from 'src/components/commons'
 import { useDeployContext } from 'src/contexts'
+import { PageHeader } from '../PageHeader'
 
-export default function ReviewStep() {
-  return <DeployButton label="Deploy" />
-}
-
-function DeployButton({ label }: { label: string }) {
+//TODO: Allow to review the configuration before deploying
+export function ReviewStep({ togglePreviousStep, toggleNextStep }) {
   const { deploy } = useDeployContext()
-  function handleDeploy(e: React.SyntheticEvent) {
-    e.preventDefault()
+
+  function handleOnClickNext() {
     deploy()
   }
   return (
     <div>
-      <Button
-        type="button"
-        onClick={handleDeploy}
-        className="font-bold text-xl"
-      >
-        {label}
-      </Button>
+      <PageHeader title="Preview" />
+      <p>TODO</p>
+      <div className="flex justify-between mt-6">
+        <Button
+          onClick={togglePreviousStep}
+          variant="outlined"
+          className="flex items-center gap-1"
+        >
+          <ChevronLeftIcon className="h-5 w-5" />
+          <span>Edit Configuration</span>
+        </Button>
+        <Button
+          onClick={handleOnClickNext}
+          variant="filled"
+          className="flex items-center gap-1"
+        >
+          <span>Deploy</span>
+          <ChevronRightIcon className="h-5 w-5" />
+        </Button>
+      </div>
     </div>
   )
 }
